M4 Wave 2 B2: strict flags for adew, apps, apt/suite, basics/{common,x,lib}, syntax/{tlex,parse,sym}, utils; fixes 5 real LP64/signal-handler bugs
FossilOrigin-Name: 2b2e570ca739e6d8574ba9d3d297ed12effe576cc2d32d4af1edc08489fd5e6c

+++ src/atk/basics/x/atoms.c
@@ -69,11 +69,11 @@ static Atom * SetAtoms(Display *dpy, char *buf)
 	return NULL;
     }
     p=buf;
-    sprintf(p, "%d ", NUMATOMS);
+    sprintf(p, "%lu ", (unsigned long)NUMATOMS);
     p+=strlen(p);
     for(i=0;i<NUMATOMS;i++) {
 	result[i]=XInternAtom(dpy, atomnames[i], FALSE);
-	sprintf(p, "%s %d ", atomnames[i], result[i]);
+	sprintf(p, "%s %lu ", atomnames[i], (unsigned long)result[i]);
 	p+=strlen(p);
     }
     XChangeProperty(dpy, RootWindow(dpy, DefaultScreen(dpy)), xim_ATOMS, XA_STRING, 8, PropModeReplace, (unsigned char*)buf, strlen(buf));
